订单号统一改为long型

+++ account-all/account-debt/src/main/java/com/atzuche/order/vo/req/AccountInsertDebtReqVO.java
@@ -18,7 +18,7 @@ public class AccountInsertDebtReqVO {
     /**
      * 主订单号
      */
-    private Integer orderNo;
+    private Long orderNo;
     /**
      * 租客子订单号
      */
