+++ account-all/account-renter-rentcost/src/main/java/com/atzuche/order/entity/AccountRenterCostSettleDetailEntity.java
@@ -23,7 +23,7 @@ public class AccountRenterCostSettleDetailEntity implements Serializable {
 	/**
 	 * 主订单号
 	 */
-	private Integer orderNo;
+	private Long orderNo;
 	/**
 	 * 子订单号
 	 */
